# docs/openapi-v2.yaml
openapi: 3.0.3
info:
  title: Daliil Ay Khidma API
  version: 2.0.0
  description: RESTful API for Business Directory Platform
  contact:
    name: API Support
    email: support@daliilaaykhidma.com
  license:
    name: Proprietary
paths:
  /api/v2/auth/login/:
    post:
      operationId: auth_login_create
      description: Custom token obtain view
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomTokenObtainPairRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CustomTokenObtainPairRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CustomTokenObtainPairRequest'
        required: true
      responses:
        '200':
          description: No response body
  /api/v2/auth/refresh/:
    post:
      operationId: auth_refresh_create
      description: |-
        Takes a refresh type JSON web token and returns an access type JSON web
        token if the refresh token is valid.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
...
  /api/v2/auth/profile/:
    get:
      operationId: auth_profile_retrieve
      description: Get current user profile
      tags:
      - auth
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v2/auth/logout/:
    post:
      operationId: auth_logout_create
      description: إبطال Refresh Token عند تسجيل الخروج من تطبيق الموبايل.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v2/auth/password-reset/:
    post:
      operationId: auth_password_reset_create
      description: إرسال رابط الاستعادة دون كشف ما إذا كان البريد مسجلاً.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v2/auth/password-reset/confirm/:
    post:
      operationId: auth_password_reset_confirm_create
      description: التحقق من رمز الاستعادة وحفظ كلمة المرور الجديدة.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v2/home/:
    get:
      operationId: home_retrieve
      description: Return all initial home widgets in one request.
      tags:
      - home
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v2/app-config/:
    get:
      operationId: app_config_retrieve
      tags:
      - app-config
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v2/admin/notifications/send/:
    post:
      operationId: admin_notifications_send_create
      tags:
      - admin
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v2/admin/users/:
    get:
      operationId: admin_users_list
      parameters:
      - in: query
        name: is_active
        schema:
          type: boolean
      - in: query
        name: is_staff
        schema:
          type: boolean
      - in: query
        name: is_superuser
        schema:
          type: boolean
      - name: ordering
        required: false
        in: query
        description: أي حقل يجب استخدامه عند ترتيب النتائج.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: رقم الصفحة ضمن النتائج المقسمة.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: عدد النتائج التي يجب إرجاعها في كل صفحة.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: مصطلح البحث.
        schema:
          type: string
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminUserList'
          description: ''
    post:
      operationId: admin_users_create
      tags:
      - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
  /api/v2/admin/users/{id}/:
    delete:
      operationId: admin_users_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: نوع رقم صحيح فريد يحدد هذا المستخدم.
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No response body
    get:
      operationId: admin_users_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: نوع رقم صحيح فريد يحدد هذا المستخدم.
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
    patch:
      operationId: admin_users_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: نوع رقم صحيح فريد يحدد هذا المستخدم.
        required: true
      tags:
      - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAdminUserRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAdminUserRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAdminUserRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
    put:
      operationId: admin_users_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: نوع رقم صحيح فريد يحدد هذا المستخدم.
        required: true
      tags:
      - admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AdminUserRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
          description: ''
...
  /api/v2/business-owner/dashboard/stats/:
    get:
      operationId: business_owner_dashboard_stats_retrieve
      description: Business owner dashboard
      tags:
      - business-owner
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessOwnerStats'
          description: ''
...
  /api/v2/businesses/nearby/:
    get:
      operationId: businesses_nearby_retrieve
      description: Return the nearest verified businesses within a radius.
      tags:
      - businesses
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessDetail'
          description: ''
...
  /api/v2/favorites/toggle/:
    post:
      operationId: favorites_toggle_create
      description: Favorite ViewSet
      tags:
      - favorites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/FavoriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FavoriteRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
          description: ''
...
  /api/v2/deals/{slug}/claim/:
    post:
      operationId: deals_claim_create
      description: Claim a deal
      parameters:
      - in: path
        name: slug
        schema:
          type: string
          title: URL Slug
        required: true
      tags:
      - deals
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealDetailRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DealDetailRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DealDetailRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealDetail'
          description: ''
...
  /api/v2/reviews/:
    get:
      operationId: reviews_list
      description: Review ViewSet for public API
      parameters:
      - in: query
        name: business
        schema:
          type: integer
      - in: query
        name: is_approved
        schema:
          type: boolean
      - name: ordering
        required: false
        in: query
        description: أي حقل يجب استخدامه عند ترتيب النتائج.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: رقم الصفحة ضمن النتائج المقسمة.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: عدد النتائج التي يجب إرجاعها في كل صفحة.
        schema:
          type: integer
      - in: query
        name: rating
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: مصطلح البحث.
        schema:
          type: string
      tags:
      - reviews
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReviewList'
          description: ''
    post:
      operationId: reviews_create
      description: Review ViewSet for public API
      tags:
      - reviews
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReviewCreateRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewCreate'
          description: ''
...
  /api/v2/reviews/{id}/like/:
    post:
      operationId: reviews_like_create
      description: Toggle a like without allowing users to like their own review.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: نوع رقم صحيح فريد يحدد هذا Review.
        required: true
      tags:
      - reviews
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
          description: ''
...
  /api/v2/notifications/read-all/:
    post:
      operationId: notifications_read_all_create
      tags:
      - notifications
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
          description: ''
...
components:
  schemas:
...
    CustomTokenObtainPairRequest:
      type: object
      description: Custom token serializer with user info
      properties:
        username:
          type: string
          writeOnly: true
          minLength: 1
        password:
          type: string
          writeOnly: true
          minLength: 1
      required:
      - password
      - username
...
    PaginatedAdminUserList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
...
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
      required:
      - access
      - refresh
    TokenRefreshRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...